requirements.txt
torch>=1.4.0
torchvision>=0.5.0
dominate>=2.4.0
visdom>=0.1.8.8
wandb~=0.16.6

numpy~=1.23.5
pillow~=10.3.0
scikit-image~=0.19.3
requests~=2.31.0
scipy~=1.4.1
opencv-python~=4.9.0.80